arcappcompat: Automate Clamshell resize testcase with resize lock feature.
Clamshell: Resize test case with resize lock feature
verifies that on clicking the dropdown menu the user
should be able to select the resizable option. On
selecting the resizable option, the app should be
resizable.

BUG=b:197014560
TEST=tast run <helios> arcappcompat.Gmail.vm_clamshell_mode

Checked the test case for all arcappcompat app tests. The test case is working as intended.

Change-Id: I439596df3789b1352ddf8042a30eb9102c1f9fe7