tast-tests: Ensure MyFiles mount in tests that use ARCVM Download/
ARCVM's Downloads integration depends on the MyFiles volume mounted on
the Android side. However, currently, there are Tast tests that depend
on the Downloads integration but do not check the MyFiles mount status.

These tests should check the MyFiles mount status before accessing
ARCVM's Download/. It will reduce their (potential) flakiness.

Also, this change will allow us to easily specify the cause of a test
failure when it is caused by the ARC MyFiles volume missing. Currently,
error messages for such failures depends on how the tests try to access
ARCVM's Download/ and do not provide direct insights for the cause.

In the course of this change, we have added a new file
chromiumos/tast/local/arc/volume.go. The definitions of functions to
wait for ARC volumes to be mounted are moved to the file, so that the
functions can be called even from tests outside the arc package. Also,
some constants related to ARC volumes are moved to the file, too.

BUG=b:198992523
TEST=tast run <DUT> arc.DownloadManager* arc.Downloads*
TEST=arc.DownloadsFolder* arc.Drivefs* arc.EnableExternalStorage*
TEST=arc.MyFiles* arc.RemovableMedia* camera.CCAUIIntent*

Change-Id: I4e3d94aa04526650649f5adde3bed736c45ab8ad