Get ui job status from upstart in ui.SessionManagerRespawn
Switch the test to query upstart for the ui init job's status as
opposed to looking at pids. This allows to correctly distinguish all
relevant cases, such as whether the previous instance is still in the
process of being stopped vs. the respawn logic actually having decided
to not respwan.

Relatedly, the test previously used a relatively small timeout of 5
seconds after which it decided that a respawn didn't take place. 5
seconds was observed to be insufficient in same cases. However, since
test success needs to wait for respawn to stop, the timeout directly
contributed to test latency so raising it would have been
problematic.. With the new way of determining status, the test can
detect the relevant situation directly, so it is safe to raise the
timeout without adverse effects on test execution time.

In combination, these two changes should resolve the observed test
flakiness.

BUG=b:197796152
TEST=ui.SessionManagerRespawn passes in betty VM, no flakes in 100 runs.

Change-Id: I8413be61a806b5c4b0cb6534a3daba79105793c4