tast-tests: Promote local/action package to common
It is not allowed for "common" code to depend on "local" code.
The use of "local/action" package in "common/bond" package breaks
the rule. The "local/action" package is generic and could be
useful to code outside "local". The CL promotes it to "common"
to fix the broken deps rule.

BUG=b:199110420
TEST=tast run $DUT ui.MeetCUJ.16p

Change-Id: I181356c60c6c91e4a2cdac47bcf5a9a6aa80da97