webrtc: add captureFromElement test and "canvas" test case
This CL adds a webrtc.CaptureFromElement test to verify the W3C
"Media Capture from DOM Elements" spec [1], a.k.a. as captureStream()
and used for e.g. Meet post-effects video capture. The spec has
two parts: one for <canvas> and one for <video>; this CL only
covers the former.

[1] https://w3c.github.io/mediacapture-fromelement/

BUG=b:199174572
TEST=tast run <kohaku IP> webrtc.CaptureFromElement.* works

Change-Id: I688e84add73d9edf3f15e0570c66d42886f40d88