boot_mode/pre: Allow booting in usb_dev mode
This CL adds an pre.UsbDevMode() precondition which allows booting
the DUT in usb_dev mode where it can load a test image from USB
device and operate from it, for example in tests invoking the
chromeos-install and other internal disk operations.

This CL is mostly focused on two parts of that process:
* Setting GBB flag FORCE_DEV_BOOT_USB flag which allows the firmware
  to boot from USB device in developer mode
* Pressing the Ctrl-U key combination at boot to actually proceed into
  booting straight from USB device

There's also a change in reporters.CurrentBootMode() to make the new
boot mode discoverable by checking if current boot device is removable.

TEST=firmware.EMMCWriteload

Change-Id: I0b7a0a42b6ac1de86ab2e42515aee48711a3f3a3