wifi: bss_tm_request: start property watcher later
In the waitForScan=true case, the scan ends up taking a significant
portion of the timeout, some times so much that it leaves less than
enough time for the connection to actually happen. Move the start of the
property watcher after we bring up the second AP and discover it from
the DUT. Note that in doing this, we are also more certain that the roam
happens from the BSS TM request, as opposed to by "natural causes".

BUG=b:197804708
TEST=run test several times, no failures

Change-Id: I895bca7c61cb483c2ab9a8a525a62527770218ca