kernel.ConfigVerify: Confirm strong stack protector and per-task canaries
We want to make sure that we're using the strong stack protector config,
not just the normal version. Similarly, we want to confirm that we're
using the per-task stack canary on arm64 devices, so that each task gets
a different canary instead of the same one calculated at boot.

BUG=None
TEST=tast run $DUT kernel.ConfigVerify

Change-Id: Iaf6df174d23828c6787225413a529307bb92a4d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3146592
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>