Files app: Fix Tast tests interacting with context menu
The pseudo-elements ::before and ::after have been added to the a11y
tree in crbug.com/1199303.  This broke the Tast tests that interact with
menu items that have a keyboard shortcut, because the keyboard shortcut
is presented in a ::after pseudo-element.

This CL fixes the tests by adding the keyboard shortcut to the Tast
tests.

BUG=chromium:1247077, chromium:1247076, chromium:1199303
TEST=tast run filemanager.Smoke filemanager.ImageQuickView

Change-Id: I730ba541f0e7dbf1db1a6d944416ffb87bfcecbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3146695
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>