tast: trogdor: Skip wifi.SimpleConnect.8021xwep
It was reported that on R92.1 the test case 8021xwep on trogdor failed
due to high ping packet loss rate. After investigation the FW 746 drop
may be the culprit. According description from Qualcomm, the changes
which potentially cause the regression are security fixes.

While we are following up the security fixes, skip the test case for
now on trogdor and re-enable it after new FW lands and gets verified.
Also skip it on strongbad which uses the same WiFi chipset as trogdor.

BUG=b:194644867
TEST=tast run -var="router=localhost:2210" -var="pcap=localhost:2212"
localhost:2200 wifi.SimpleConnect.8021xwep, and then message "PlatformId
matched with skip-on list" shows up only on trogdor and strongbad. CQ
won't break.

Change-Id: Ibb51259abb073cae184af3d5e402ffb0b6ef35d2