tast-tests: fix activating input field
A few bugs reported recently around physical keyboard tests (even I could not repro it locally). The same root cause is after mouse clicking a field, failed to wait for the field to be activated. It causes flakiness of tests but not relevant to the scenario.

In the current implementation, it clicks input field once and wait unlimited time for the field to be active. If the local server has a render issue or any problem that it is not ready to receive click event, then it fails until test timeout (5 mins).

In this fix:
1. Shorten the timeout of each wait to 3 seconds after click.
2. Add retry to click the field again if it is not activated due to surprises.

BUG=b:199200234
TEST=tast run <DUT> inputs.PhysicalKeyboardEnglishTyping*

Change-Id: Ice9f71f5959a42242598fe87fb951af3b12bd0f8