arc: Add constants for WM phone and tablet size apps to WM helper
These apps will be more widely used, so they shouldn't be defined
in individual tests.

Also, the package name of "WM Secondary" is a bit confusing as it
dooesn't mention that it's allowlisted to be launched in maximzied.
So this CL removes the usage of it.

BUG=b:194156232
TEST=arc.WindowDefaultBoundsAllowlist.vm

Change-Id: I91ba652d50100785adb2526374dac0f91840de9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3149879
Tested-by: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>