tast-tests: Remove *PIPEnergyAndPower.*_blend*
This CL removes *PIPEnergyAndPower.*_blend* which put things in front of
a PIP video to ensure that if it is promoted to an overlay, it will use
the underlay overlay strategy. Now that the PipRoundedCorners feature is
enabled by default, any PIP video promoted to an overlay will use the
underlay overlay strategy anyway. *PIPRoundedCornersUnderlay* verify
this fact. Thus, *PIPEnergyAndPower.*_blend* have become redundant with
the corresponding tests without "blend" in their names.

BUG=chromium:1093933
TEST=tast run <DUT> *PIPEnergyAndPower.*

Change-Id: I2acf77b9440c69e599c6ba928eb5bcdacb2084b3