cuj: adjust loading status check for websites
When running everyday multitasking CUJ, it has been seen that content
sites sometimes take minutes to reach quiescence. When this situation
occurred, it can be seen from screenshots that the UI has loaded but
background tasks prevented the site to reach quiescence. This commit
tries to resolve the issue by adjusting the loading status check to
verify the document state is in either 'complete' or 'interactive'
state.

BUG=b:200147496
TEST=tast run <vars> <dut> ui.EverydayMultiTaskingCUJ.*

Change-Id: Ia207e7d78288d73b6d4a9b371c565a546cbaa50c