multidevice: add multidevice setup fixture and utilities
Add fixtures and utilities for multidevice tests (phone hub, smart lock,
etc.). One fixture is for Android setup, which handles GAIA login,
settings+logging configuration, and installing the Multidevice Snippet.
The other is for logging in to CrOS and pairing the devices. Many of the
setup steps are shared with the Nearby Share fixtures, so some common
code has been copied to the multidevice package. I will refactor the
Nearby fixtures later (blocked by b/199110070, since many of the changes
are in the common nearbyshare package).

BUG=b:187793760
TEST=tast run <dut> multidevice.PhoneHubDoNotDisturb

Change-Id: I7e48e52f1e4a35fc5e768124494deef964b6cabe