phonehub: Add Phone Hub "Silence Phone" test
Add an automated test for Phone Hub's Do Not Disturb/Silence Phone pod.
The test toggles the Silence Phone pod from Phone Hub, and checks that
the Do Not Disturb setting on the Phone is updated accordingly. It also
checks the reverse direction, that the Phone Hub pod's status is updated
based on the phone's DND status.

BUG=b:187793760
TEST=tast run <dut> phonehub.SilencePhone

Change-Id: I277d1a4778fc306a8b135bc5f9daa95090398955
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3169587
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>