multidevice: Add Phone Hub Do Not Disturb test
Add an automated test for Phone Hub's Do Not Disturb/Silence Phone pod.
The test toggles the Silence Phone pod from Phone Hub, and checks that
the Do Not Disturb setting on the Phone is updated accordingly. It also
checks the reverse direction, that the Phone Hub pod's status is updated
based on the phone's DND status.

BUG=b:187793760
TEST=tast run <dut> multidevicePhoneHubDoNotDisturb

Change-Id: I277d1a4778fc306a8b135bc5f9daa95090398955