tast-tests: Snap about:blank in ui.OverviewPerf
The present CL is for investigating why tablet split view degrades the
performance of the overview enter/exit animations. One of the windows in
ui.OverviewPerf will be about:blank instead of a computationally
expensive animation. That window is the one that will be snapped for the
split view cases. This CL is not intended to be used with
ui.OverviewPerf.lacros.

BUG=chromium:1124599
TEST=manual

Change-Id: I634f7ef12d3908df9988531f004119f6e04d8a9a