tast-tests: Include about:blank but snap animation in ui.OverviewPerf
The present CL is for investigating why tablet split view degrades the
performance of the overview enter/exit animations. One of the windows in
ui.OverviewPerf will be about:blank instead of a computationally
expensive animation. However, a window with the computationally
expensive animation will be snapped for the split view cases. This CL is
not intended to be used with ui.OverviewPerf.lacros.

BUG=chromium:1124599
TEST=manual

Change-Id: Ic3057b5b5361819417f7bd2ba13e9f234b6e5c11