Tast: Relax test wifi.FunctionalAfterCSA.router
The DUT sometimes auto-connects after deauthenticating the client on the
router which casues the function AssureDisconnect() to fail. The test
goal is to make sure that DUT can connect to the router after CSA followed
by a disconnect event. Auto-connecting shows that the DUT can connect to
the AP and the test should not fail in this case.

Relaxing the test by not failing in this case and forcing a client
disconnection for the next iteration connection to work.

BUG=b:197414763
TEST=ran tast run $DUT wifi.FunctionalAfterCSA.router

Change-Id: I44381880409b95f6618c5e1a96cb5b1dfbaad096