tast-tests: fix PhysicalKeyboardEmoji
A few bugs reported recently around physical keyboard tests (even I could not repro it locally). The same root cause is after mouse clicking a field, failed to wait for the field to be activated. It causes flakiness of tests but not relevant to the scenario.

Left click has been workaround with a retry in https://crrev.com/c/3151091. This cl is to fix right click.

BUG=b:199200234
TEST=tast run <DUT> inputs.PhysicalKeyboardEmoji*

Change-Id: Ic5d88e728281162544c61cecfb53b89178521064