CrOS AppList: Fix folder related test failures
Remove fixtures from CreateAndFillFolder and RemoveAppsFromFolder and
instead create the fake apps and call chrome.New within the tests
themselves. This ensures that the test's session has enough apps for
the foldering actions that need to be performed. This also makes sure
that performance tests that rely on the same fixtures do not have their
stability impacted by the reduced number of pages/apps available.

Changes the way CreateFolder() works by finding the first two non-folder
items in the launcer and dragging one on top of the other to create the
folder. This change is useful in the case where one of the items are a
default created folder.

BUG=b:190654657
TEST=Ran all foldering tests locally

Change-Id: Ide3544edaecb6d68c69a32b3948bfb0deaec8481
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3163794
Tested-by: Matthew Mourgos <mmourgos@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Tast-Review: Kyle Shimabukuro <kyleshima@chromium.org>