CrOS AppList: Ensure CreateFolder() works by using non-folders
Changes the way CreateFolder() works by finding the first two non-folder
items in the launcer and dragging one on top of the other to create the
folder. This change is necessary because previously, the first or second
available item could be an already existing folder. This meant that a
new folder would not be created, possibly causing the test to fail.

Also rename the folder which gets filled in create_and_fill_folder.go
so that the folder is not reused in subsequent tests.

BUG=b:190654657
TEST=Ran all foldering tests locally

Change-Id: Ide3544edaecb6d68c69a32b3948bfb0deaec8481