WaitForCrashFiles: Don't poll-break on unreadable meta files
There's a known issue with cryptohome 'flickering' occasionally.
(b/189707927) If one process writes a file, a different process
trying to read it the instant the file shows up may not be able
to. So don't testing.PollBreak if the file can't be read, just
retry.

BUG=chromium:1246250
Fixed: chromium:1246250
TEST=tast -verbose run --extrauseflags chrome_internal localhost:2222 ui.ChromeCrashLoggedIn.*

Change-Id: I27d434bfcce268b1c73d383f87f1985f94b8570a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3166053
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>