tast-tests: Add variant of ChromeVox hint test
Previously, only one variant of the ChromeVox hint test existed. In this
single variant, we waited for the ChromeVox hint dialog to appear, canceled
the dialog, then waited for a spoken announcement.

This change adds another variant of the test where we accept the dialog,
instead of dismiss it. This ensures that both user journeys of this feature
are tested.

Lastly, since this change introduces a new test, we demote this test
from critical to informational. Once this test is proven stable, we will
promote it back to critical.

BUG=chromium:1159107
TEST=tast -verbose run 100.90.28.180 a11y.ChromevoxHint.*

Change-Id: I2b49edae6dde0351d36f229a21ce63fe16f1ac10