arcappcompat: Automate Clamshell fullscreen test case with resize lock feature.
Clamshell: Fullscreen test case with resize lock feature verifies
that on clicking the dropdown menu the user should be able to select the
resizable option. On selecting the resizable option, the app should be
resizable. On using ash.WMEventFullscreen api, the app should be in
full screen mode without any crashes.

BUG=b:197014750
TEST=tast run <helios> arcappcompat.Gmail.vm_clamshell_mode

Checked the test case for all arcappcompat app tests. The test case is working as intended.

Cq-Depend: chromium:3141971
Change-Id: I46c912dcf23fb5f717a07bbd9fe645db909ee05b