multidevice: add wrapper for Multidevice Snippet
Add a wrapper in Tast for the Multidevice Mobly Snippet. This snippet
provides an RPC for pairing an Android phone with a Chromebook,
bypassing the multidevice setup flow and enabling use of the "Connected
devices" features such as Phone Hub, Smart Lock, WiFi Sync, etc. The
wrapper also includes some basic utilities for logging, and wraps the
ADB commands used to interact with the phone's Do Not Disturb setting,
since we will test Phone Hub's interaction with this setting.

BUG=b:187793760
TEST=tast run <dut> multidevice.PhoneHubDoNotDisturb

Change-Id: Ia087fc3ac94ba4a4e67a81fcc46621c1da220622