wifi.SetTXPower: Correctly validate static devices
Static devices are devices that do not mode switch from tablet to
non-tablet, but will require setting WiFi transmit power levels. Work
was added to support static devices
(https://chromium-review.googlesource.com/q/topic:powerd-static-wifi-transmit-power-initialization),
but the test was not updated to accommodate static device behavior.

Spherion, for example, supports non-tablet mode WiFi transmit power mode
statically and fails when the test attempts to assert tablet mode power
settings. This updates the test to see if the device is static, and if
so validates that WiFi transmit power levels are set correctly for the
configured static mode.

BUG=b:200622521
TEST=`tast run ${DUT} wifi.SetTXPower` succeeds on Spherion.

Change-Id: Ief76f7790c98e97acae55f1046d86ec5fabfd7f1