webrtc: fix CaptureFromElement.* flakes
CaptureFromElement.* check if the captured-from-canvas video is
not all black pixels. This is flaky because we use OffscreenCanvas
drawImage() and getImageData() that are not synchronous, hence
being racy. This CL adds logic to try looking for black pixels a
number of times, with a 1/60Hz pauses in between.

BUG=b:199174572
TEST=while true; do tast run <volteer> webrtc.CaptureFromElement.canvas* | grep "Z webrtc.CaptureFromElement.canvas"; done;
is not flaky anymore.

Change-Id: I3b2455086a40780425c918a5c730a4d5c9c04d50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3174557
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tast-Review: Po-Hsien Wang <pwang@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>