cuj: tab switch - adjust loading status check for websites
When running tab switch CUJ, it has been seen that content sites
sometimes take minutes to reach quiescence. When this situation
occurred, it can be seen from screenshots that the UI has loaded but
background tasks prevented the site to reach quiescence. This commit
tries to resolve the issue by adjusting the loading status check to
verify the document state is either 'complete' or 'interactive' state.

BUG=b:200147496
TEST=tast run <vars> <dut> ui.TabSwitchCUJ2.*

Change-Id: Ia88ae3affe0d9c2260f28bb1f9b7b8559f4cccb6