lacros-au: Add remote test with skeleton service
This change adds the initial remote autoupdate test that verifies
the Lacros update path of Rootfs => Stateful.
The test will connect to TLS for provisioning, to a test service
on a DUT for test execution.
Currently, the test service is just created with skeleton code.
It will be implemented in a follow-up CL.

See https://sequencediagram.googleplex.com/view/6215323990622208
for the workflow and go/lacros-autoupdate-test-proposal.

BUG=chromium:1247311
TEST=fast_build.sh -T (unittest)
TEST=# Start TLW + TLS locally
TEST=tast run -var=provision.TLSAddr=${ADDR} ${DUT} lacros.UpdateRootfsToStateful

Change-Id: I81f05415226aede1ee006fb39d9413e2eb7ab06b