check_disk_space: only require 24MiB of margin
The size of the OS has grown on some boards (e.g. coral) and they're
getting close to the hard limit, with less than 32MiB still available on
the rootfs.
Let's make this test a bit more permissive (minimum 24MiB of free space
instead of 32MiB) temporarily. That avoids blocking releases while
giving a little bit of time to optimize the size of the OS.

BUG=b:200746365
TEST=./platform/tast/fast_build.sh -T

Change-Id: Ie0cfa9d66add731cc1a836a09e2d63daa0b5fd98