tast-tests: Add a test for 'Check Network' button in optin dialog
We've added a new button to optin dialog that displays when there is a
network related error during optin. In this test we disable ARC
networking interfaces and initiate optin to make it fail due to network
failure and then we validate that the button is shown and clicking the
button launches the Diagnostics app.

BUG=b:200176336
TEST=tast run DUT arc.OptinNetworkError.vm

Change-Id: Ie86509f1c0819906bf42bacaa9ec8c3cd456d57a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3176772
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Tast-Review: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>