video: better logging PlatformEncoding
PlatformEncoding concatenates the execution of a few commands,
in particular it encodes the video, then it decodes it again
(to compare it with the original using tiny_ssim). This CL
adds file logging for the "decoder" section. Since the call
for decoding happens together with the tiny_ssim comparison
call, it renames a few variables.

BUG=b:198199171
TEST=tast run <kevin> video.PlatformEncoding.v4l2_h264_*

Change-Id: Ia0e152faa3475bfff1ef79a817baa2483a86ed8f