Fix flakiness of LaunchSystemWebAppsFromURL
Sometimes the browser is closed after opening a SWA, which causes the
failure of the following test. Change to new a connection at the
beginning of each subtest.
The previous screen record sometime does not work with some error,
change it to record from keyboard.

BUG=None
TEST=Ran on hatch

Change-Id: I8a355052363c455ef5c2de350a0e852fa28f65b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3177045
Tested-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Judy Wu <jinrongwu@google.com>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>