A test to investigate "Too many open files" issue of virtio-fs
A lot of file descriptors seem to be left open when Files.walkFileTree
or Files.walk is performed on a directory exported by virtio-fs.
This test is to locally reproduce the issue.
arc.TooManyOpenFiles.sdcard creates a lot of files in Android's
/storage/emulated/0/Documents/ArcTooManyOpenFiles, and launch an Android
app. You can trigger Files.walk and Files.walkFileTree by pressing the
corresponding button for each method.
arc.TooManyOpenFiles.myfiles does the same thing for
/storage/<MyFiles-UUID>/ArcTooManyOpenFiles. When virtio-fs is used on
MyFiles, pressing any of the buttons causes "Too many open files"
exception when numberOfFiles is large enough (~10,000). This is because
mount-passthrough sets ulimit -n 8192.

BUG=b:182870415
TEST=(cr) $ cros_workon --board=${BOARD} start tast-local-apks-cros
TEST=(cr) $ emerge-${BOARD} tast-local-apks-cros
TEST=(cr) $ cros deploy --board=${BOARD} --root=/usr/local <DUT> \
TEST=       tast-local-apks-cros
TEST=(cr) $ tast run <DUT> arc.TooManyOpenFiles.sdcard
TEST=$ ssh <DUT>
TEST=# find /proc -name status -exec grep -H virtio-fs {} \;
TEST=# strace -f -p ${PID-fs} -o ${file-1}
TEST=Press the "FILES.WALKFILETREE" button on the app window.
TEST=Wait for a while until Files.walkFileTree is finished.
TEST=# lsof -p ${PID-fs} > ${file-2}
TEST=(cr) $ tast run <DUT> arc.TooManyOpenFiles.myfiles
TEST=# find /proc -name status -exec grep -H virtio_9p {} \;
TEST=# strace -f -p ${PID-9p} -o ${file-3}
TEST=Press the "FILES.WALKFILETREE" button on the app window.
TEST=# lsof -p ${PID-9p} > ${file-4}

Change-Id: Ie3d0aec1c78814027ffee74ef49e147ce4b9c2e3