video/data: Add javascript code of demuxing MP4 container
This adds mp4_demuxer.js that makes use of mp4box js. It enables
getting a ebitstream data that represents a single video frame.
The license of mp4box js is BSD 3-Clause [1]. The license of
mp4_demuxer.js is Chromium BSD, although it is put in webcodecs/samples [2],
because it is written by tguilbert@chromium.org.

[1] https://github.com/gpac/mp4box.js
[2] https://github.com/w3c/webcodecs/blob/0dd2a71f04a806992f4b7e11c78cd2439b374dd4/samples/mp4-decode/mp4_demuxer.js

BUG=b:196307009
TEST=None

Change-Id: I73db1e2e28fa4ba53dae400284b161054a953ea2