video: Add encoder test in Webcodecs API
This CL introduces the first test to verify WebCodecs.
WebCodecs API enables a web app to encode a video frame using
native encoders in Chrome. In the test, we encode the frames by
 software or hardware encoder using WebCodecs API and verifies
the encoded bitstream by PSNR and SSIM.

The webcodecs javascript code is inspired by webcodecs_common.js
in chromium [1].

[1] https://source.chromium.org/chromium/chromium/src/+/main:content/test/data/gpu/webcodecs/webcodecs_common.js

BUG=b:196307009
TEST=video.WebcodecsEncode.*

Change-Id: I57535af913a57fab62ab0ba9bc244eb636ca5fe7