network: DNSProxy: Wait for the proxy settings to be propagated
DNS queries through the proxy sometimes are successful even if
the network is blocked. This might be caused because the
settings or the network configuration is not updated yet.

This patch is an experiment to test if the failures are
because of the propagation issue. If this fixes the failures,
update the arbitrary timeout to DBus polling instead. Otherwise,
remove the timeout from the test.

BUG=b:199093160
TEST=tast run ${DUT} network.DNSProxy.*
TEST=tast run ${DUT} network.DNSProxyOverVPN.*

Change-Id: I76bd09ee77ea600b029452163a8d0f4ae9647fe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3184845
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Tast-Review: Po-Hsien Wang <pwang@chromium.org>