camera: Adding action_sequence to CCAUIStress test
This CL adds action_sequences var to CCAUIStress that allows a list of actions to be stressed in a single iteration.
It also adds the table mode option.

BUG=b:195902740

E.g., On DUT's that support dual camera, this action_sequence will
capture the images using both front and back cameras for the given
iterations.

TEST=tast run -var=action_sequence=take-photo,switch-camera,take-photo -var="iterations=5" <DUT_IP> camera.CCAUIStress.real

Change-Id: I486643c8f0b9b97929a314d9fc5aaf5bb4a71a00
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3185451
Reviewed-by: Ambalavanan M M <ambalavanan.m.m@intel.com>
Tested-by: Ambalavanan M M <ambalavanan.m.m@intel.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Alex Levin <levinale@google.com>