Addressed flakiness when TAST test performs GAIA signin.
The ChromeOS OOBE hosts multiple GAIA WebView targets at one time. Only
one of these targets are visible and can be used for enrollment or
signin.

Until this change, there is confusion over which WebView to use, which
in turn results in test flakiness, especially at the step between when
test completes enterprise enrollment and when test sign in as the
enrolled user.

This change addresses this issue by emulating the actions of the
Catapult telemetry project's oobe.py script, and identifies the correct
GAIA WebView target by looking for a <base> element in the WebView. The
change was suggested by ChromeOS OOBE dev owners.

Also attempted to break long lines in enroll.go and gaia.go for the
changes I have added recently.

BUG=chromium:1194164
TEST=ran a Gaia signin test and an enterprise enrollment test on a physical device.

Change-Id: I5d48bcd35f7e1379125565e938295b7657ee52e2