tast-test: AutoHideSmoke should reset shelf state at the end
AutoHideSmoke test was changing shelf auto-hide state, without later
resetting it to the original state, at the same time expecting the shelf
to be in a particular state (shelf always shown). This meant that when
AutoHideSmoke.clamshell and AutoHideSmoke.tablet_mode were run
consecutively, the latter run would fail to find an option to auto hide
shelf (given that the shelf was already autohidden).

This change makes the test more robust by having test update the shelf
state to the expected one when it starts, but also adds deferred call to
change shelf auto-hide state to the one found at the beginning of the
test.

BUG=b:188196863
TEST=ran AutoHideSmoke* locally

Change-Id: I02ea358de16a046102e6740ad0fec9e8ab416ed3