Relax display zoom change bounds check
Because of our complicated floating point rounding logic, the
actual bounds can be 3PX different from the expected one, but the
actual bounds is the one we want, so it's working fine.

As it's virtually impossible to calculate the exact value in Tast,
this CL just relaxes the condition of the bounds check. Even with
this CL, the bounds check works to catch regressions in most cases.

BUG=b:194156232
TEST=arc.NonResizableClamshell

Change-Id: I21b4002f980b6cc9033ebf7bf3c7f0d55827a86d