arc: Only check display orientation in a binary way in WMNonresizableTablet
We've already merged a similar CL for WMResizableTablet
(crrev.com/c/3043465), but forgot to do the same for
WMNonresizableTablet, which this CL handles.

Depending on how orientation values are configured, a rotated
display doesn't necessarily have a primary orientation.

As it really depends on the configuration of the device, we need
to relax the condition.

BUG=b:194156232
TEST=arc.WMNonresizableTablet

Change-Id: I145ecaad6838aa033dc5ab7efbd0e3f3180c05ec