arc: Fix testPIPGravityQuickSettings
dispMode passed as a function parameter is supposed to be used for
some values that are consistent in the test (e.g. dsf).

On the other hand, display bounds is something can change during the
test with display rotation, so we need to query it every time we
use it.

BUG=b:194156232
TEST=arc.PIP.vm

Change-Id: Ie29f1e57ca671f37ba417faf39a7346b8fab0f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3190132
Tested-by: Kazuki Takise <takise@chromium.org>
Auto-Submit: Kazuki Takise <takise@chromium.org>
Commit-Queue: Kazuki Takise <takise@chromium.org>
Reviewed-by: Kamil Szarek <kamilszarek@google.com>