cuj: check if error is caused by Signout issue
During CUJ testing, if the sign-in account is being used too many times
to do cloud operations, account will be signed out which causes the test
case failure. For better debugging this issue, when test error is found,
we add logic to check if it is caused by signout issue and then wrap
those info in the error message.

BUG=b:201951383
TEST=tast run <vars> <dut> ui.GoogleMeetCUJ.*
TEST=tast run <vars> <dut> ui.ZoomConfCUJ.*

Change-Id: I157eef25eb081da1a05fa4c8e87d60c1f3c5c7b4