arc: Allow margin when counting PIP menu pixels
The RGB value of a gray pixel queried via the screenshot API can
be different (a lot) depending on the device.

This CL passes a margin when looking for PIP menu. The two colors
compared in the function never overlaps even with the new margin,
so this function still works perfectly.

BUG=b:194156232
TEST=arc.PIP.vm

Change-Id: I6011dc67a24a71aa1637468464bbbc1398c69bb8