health: Fix usb bus info tast
Usb device files may contain spaces. Trim the spaces to meet the healthd
output.
Previously, we joined Vendor and Product name and compared it with lsusb
directly. However, the Product name may contain spaces which is trinmed
by healthd. To also trim it in tast, compare Vendor and Product name
seperately, and trim the spaces of results from lsusb.


BUG=b:200752022
TEST=tast/fast_build -T
TEST=tast run $DUT health.ProbeBusInfo

Change-Id: I1ed8e7759d4fd1aeed28885df2423fa17b6ba0c2