blink: Add test to verify transferFromImageBitmap orientation
chromium:3171576 fixes a Blink bug which causes
ImageBitmapRenderer.transferFromImageBitmap to render images flipped
upside-down. The bug is only reproducible on CrOS devices.

Web platform tests do not support CrOS, so add a Tast test to capture
the bug behavior.

This CL depends on chromium:3171576, but it cannot be "Cq-Depend"ent
because it's in a separate LUCI project.

BUG=b:186464068
TEST=tast/fast_build -T
TEST=tast run $DUT graphics.TransferFromImageBitmapOrientation.*

Change-Id: I60ded3a4b89f2a728a4e2c4f2617b69dd05ce0c8