blink: Add test to verify transferFromImageBitmap orientation
https://crrev.com/c/3171576 fixes a Blink bug which causes
ImageBitmapRenderer.transferFromImageBitmap to render images flipped
upside-down. The bug is only reproducible on CrOS devices.

Web platform tests do not support CrOS, so add a Tast test to capture
the bug behavior.

This CL depends on https://crrev.com/c/3171576, but it cannot be
"Cq-Depend"ent because it's in a separate LUCI project.

BUG=b:186464068
TEST=tast/fast_build -T
TEST=tast run $DUT graphics.CompositorTransferFromImageBitmapOrientation.*

Change-Id: I60ded3a4b89f2a728a4e2c4f2617b69dd05ce0c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3199838
Tested-by: Andrew Wolfers <aswolfers@chromium.org>
Commit-Queue: Andrew Wolfers <aswolfers@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>