tast-tests: Make Lacros extensionArgs visible outside of Fixture
Currently only Lacros fixture uses extensionArgs to enable test
extension API on Lacros. This is not accessible outside of the fixture.

The problem is that test service that doesn't rely on fixture can't
reuse these args when it needs to pass them to Ash for Lacros.

This is a safe change not affecting any existing tests since
it just changes the visibility of the function.

BUG=chromium:1247311
TEST=tast run <DUT> lacros.ShelfLaunch

Change-Id: Ifda1b5772770e53a0a9df934023a0778dc2227bd