security: Add global seccomp policy tast test
Add a new tast test, GlobalSeccompPolicy, that checks that a specific
syscall is blocked as confirmation that the global seccomp policy is
being applied. This syscall is the LOOP_CHANGE_FD ioctl and a note is
made in the policy file of the reliance of this test on the syscall
being blocked.

BUG=b:200683781
TEST=run tast test

Cq-Depend: chromium:3194851
Change-Id: Ie44d2191ba3cf9fa2a54aa4e747f471018602275