policy: Enterprise rollback with self-built img
Add the official update server to be used for updates. Then, self-built
images do not need to manually be configured to use that server (they
have the machine they were built on as default update server).

Also touch lsb-release as it may not be created yet when the test runs
and that would cause the test to fail.

BUG=chromium:1240933
TEST=tast run -var=policy.RollbackWithGaia.confirm=ICanRollbackMyDUT
     <ip> policy.RollbackWithGaia

Change-Id: Ic841c27e5bb4523e49438cedfd103a70cc09296e