arcappcompat: adjust trackpad scroll logic
Adjusts the trackpad scroll logic to keep it within the bounds of each
model's trackpad. The scroll always starts at the center of the screen
but was using a hardcoded distance to perform the scroll previously.
This change adjusts the amount to be dynamic based on the height of the
trackpad to prevent out of bounds errors. Additionally, the time to
complete the scroll was incremented a small amount to help reduce
flakiness. The duration was derived experimentally by running on
multiple devices.

BUG=b:201668316
TEST=push apk && tast run ${HOST} arc.StandardizedTrackpadScroll*

Change-Id: I0b6e01f3963b10b9b87698b6fe6eed2e24d54ac3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3197871
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Tested-by: David Welling <davidwelling@google.com>
Commit-Queue: David Welling <davidwelling@google.com>