arcappcompat: adjust trackpad scroll logic
Adjusts the trackpad scroll logic to keep it within the bounds of each
model's trackpad. The scroll always starts at the center of the screen
but was using a hardcoded distance to perform the scroll previously.
This change adjusts the amount to be dynamic based on the height of the
trackpad to prevent out of bounds errors. Additionally, the time to
complete the scroll was incremented a small amount to help reduce
flakiness. The duration was derived experimentally by running on
multiple devices.

BUG=b:201668316
TEST=push apk && tast run ${HOST} arc.StandardizedTrackpadScroll*

Change-Id: I0b6e01f3963b10b9b87698b6fe6eed2e24d54ac3