policy: Manual rollback test with version vars
Add variables for target and source version to the manual rollback test.

The source version is required because otherwise the update request may
not set the version correctly due to either sending "ForcedUpdate" or,
if run an a self-built image, the version not being set correctly in
/etc/lsb-release.

The target version allows for deciding which image to roll back to on
runtime.

BUG=chromium:1240933
TEST=tast run -var=policy.RollbackWithGaia.confirm=ICanRollbackMyDUT
              -var=policy.RollbackWithGaia.sourceVersion=14244.0.0
	      -var=policy.RollbackWithGaia.targetVersion=14092.
	      localhost:1111 policy.RollbackWithGaia

Change-Id: I3958d34045d2ea7995cf296d165b35986cc13da7