bios/image: Add the ability to choose flashrom programmer
This CL is a simple addition to bios/image.go which allows the
tests to use different flashrom "programmers" (the -p parameter)
as the default hardcoded one is "host" (for AP firmware) and it
might be valuable for some tests to utilize the "ec" programmer
for EC firmware of course.

It currently enumerates the possible programmers as strings which
are typed as FlashromProgrammer type and it's possible to add new
entries in the future.

TEST=firmware.ECHash

Change-Id: Ie89652d07cb63eca51bbd82d714e092d5439202f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3199914
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tast-Review: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jakub Skrzypnik <js@semihalf.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>